requirements.txt
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
numpy==2.4.4
opencv-python==4.13.0.92
pillow==12.2.0
Werkzeug==3.1.8
